Atomise code
