+++ Form1.cs
@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Drawing;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices.Expando;
@@ -78,135 +77,5 @@ namespace Route_Finder
             }
         }
 
-    }
-
-    public class Node
-    {
-        private int height = 1;
-        private Dictionary<Node, int> connections = new Dictionary<Node, int>();
-        private int x, y;
-
-        public Node(int x, int y, int height)
-        {
-            this.x = x; this.y = y; this.height = height;
-        }
-
-        public int getHeight()
-        {
-            return height;
-        }
-
-        public void setHeight(int height)
-        {
-            this.height = height;
-        }
-
-        public Dictionary<Node, int> getConnections()
-        {
-            return connections;
-        }
-
-        public void setAllConnections(Node p, int dist)
-        {
-            this.connections.Add(p, dist);
-        }
-
-        public int getSpecificPoint(Node p)
-        {
-            return (int)connections[p];
-        }
-
-        public List<Node> getConnection()
-        {
-            List<Node> pointConnections = new List<Node>();
-            foreach (Node p in connections.Keys) { pointConnections.Add(p); }
-            return pointConnections;
-        }
-
-        public string getStringCoOrd()
-        {
-            return x + ":" + y;
-        }
-
-        public string getStringConnectinos()
-        {
-            String s = "";
-            foreach (Node p in getConnection())
-            {
-                s += p.getStringCoOrd() + ", ";
-            }
-            return s;
-        }
-    }
-
-    public class PointsCreation
-    {
-        private const int size = 10;
-        private List<Node> points = new List<Node>();
-        
-
-        public PointsCreation()
-        {
-
-            for (int j = 0; j < size; j++)
-            {
-                for (int ii = 0; ii < size; ii++)
-                {
-                    points.Add(new Node(j, ii, 1));
-                }
-            }
-
-            string f = System.IO.File.ReadAllText("data.csv");
-            f = f.Replace('\n', '\r');
-            string[] lines = f.Split(new char[] { '\r' }, StringSplitOptions.RemoveEmptyEntries);
-
-            int r = lines.Length;
-            int c = lines[0].Split(',').Length;
-
-            int[,] val = new int[r, c];
-
-            for (int j = 0; j < r; j++)
-            {
-                string[] line_i = lines[j].Split(',');
-                for (int jj = 0;jj< c; jj++)
-                {
-                    Int32.TryParse(line_i[jj], out val[j,jj]);
-                    Console.WriteLine(val[j,jj]);
-                }
-
-
-            }
-
-            int index = 0;
-            int i = 0;
-            foreach (Node p in points)
-            {
-                while (index < c)
-                {
-                try
-                    {
-                    Console.WriteLine(val[i, index]);
-                    p.setAllConnections(points[val[i, index]], 5);
-                    }
-                    catch (Exception e) {}
-
-                    index++;
-                }
-                i++;
-                index = 0;
-                if(i > r)
-                {
-                    break;
-                }
-            }
-        }
-
-        public Node getNode(int index)
-        {
-            return this.points[index];
-        }
-
-        public List<Node> getAllNodes() { return this.points; }
-        public int getNumberOfNodes() {  return this.points.Count; }
     }
 }
