+++ Node.cs
@@ -1,0 +1,64 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace Route_Finder
+{
+    public class Node
+    {
+        private int height = 1;
+        private Dictionary<Node, int> connections = new Dictionary<Node, int>();
+        private int x, y;
+
+        public Node(int x, int y, int height)
+        {
+            this.x = x; this.y = y; this.height = height;
+        }
+
+        public int getHeight()
+        {
+            return height;
+        }
+
+        public void setHeight(int height)
+        {
+            this.height = height;
+        }
+
+        public Dictionary<Node, int> getConnections()
+        {
+            return connections;
+        }
+
+        public void setAllConnections(Node p, int dist)
+        {
+            this.connections.Add(p, dist);
+        }
+
+        public int getSpecificPoint(Node p)
+        {
+            return (int)connections[p];
+        }
+
+        public List<Node> getConnection()
+        {
+            List<Node> pointConnections = new List<Node>();
+            foreach (Node p in connections.Keys) { pointConnections.Add(p); }
+            return pointConnections;
+        }
+
+        public string getStringCoOrd()
+        {
+            return x + ":" + y;
+        }
+
+        public string getStringConnectinos()
+        {
+            String s = "";
+            foreach (Node p in getConnection())
+            {
+                s += p.getStringCoOrd() + ", ";
+            }
+            return s;
+        }
+    }
+}
